refactor: fixed PR tasks, added deprecation notice
